+++ CHANGELOG.md
@@ -4,7 +4,17 @@ Latest first.
 
 ## STAGING
 
+### Additions
+- Add `FlagWarAPI#isUnderAttack(Nation)`
+  - Returns true if any town in the supplied nation contains a cell under attack.
+- Add `OutlawListener` and listen for the `OutlawTeleportEvent`
+  - Outlaws cannot be teleported away during an attack while in enemy territory (or enemy nation zone) during an active
+  war.
+  - Add `error.outlaw.cannot-teleport-here` translation key, sent to the outlaw.
+
 ### Misc. Changes
+- Bump Towny API dependency to 0.97.0.17; set as Min_Towny_Version
+- Bump Config Version to 1.3 (Post 0.3.0... Woopsie.)
 - Removed unnecessary deployment to GitHub Packages
 
 ## 0.3.0